Fix test directory name
