+++ Algorithms.sln
@@ -1,4 +1,4 @@
-﻿
+
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.29911.84
@@ -9,7 +9,7 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Algorithms", "src\Algorithm
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "test", "test", "{0753E0EB-74E4-4BF4-BF80-D692DA4F64E3}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Algorithm.Tests", "test\Algorithm.Tests\Algorithm.Tests.csproj", "{5171C322-2055-4C5C-88AB-E97E1068B34B}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Algorithms.Tests", "test\Algorithms.Tests\Algorithms.Tests.csproj", "{5171C322-2055-4C5C-88AB-E97E1068B34B}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{BAAACFB2-7725-4F34-BE0C-6C0280BB1578}"
 	ProjectSection(SolutionItems) = preProject
